requirements.txt
# requirements.txt

# 核心功能依赖
requests~=2.32.4
pycryptodome>=3.18.0
m3u8>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.1
lxml~=6.0.0

# 打包工具依赖
pyinstaller>=5.13.0       # 用于将Python脚本打包为可执行文件